fix: #5 - then 적용
