+++ EveryDayVoca/Controllers/FlashCardViewController.swift
@@ -51,5 +51,4 @@ final class FlashCardViewController: BaseViewController {
                           animations: nil,
                           completion: nil)
     }
-    
 }
